feat(demo): add client fallback and update demo-ui
