+++ demo/package.json
@@ -12,7 +12,7 @@
     "start": "node run-demo.js"
   },
   "dependencies": {
-    "@just-every/demo-ui": "file:../../demo-ui",
+    "@just-every/demo-ui": "^0.1.1",
     "d3": "^7.9.0",
     "dompurify": "^3.2.3",
     "dotenv": "^17.0.1",
